+++ api-scanner/src/test/resources/class-internal-annotation/java/net/corda/example/AnnotatedClass.java
@@ -1,5 +1,9 @@
 package net.corda.example;
 
+/**
+ * This class should appear in the scan output,
+ * but its @InvisibleAnnotation should not.
+ */
 @InvisibleAnnotation
 public class AnnotatedClass {
 }