+++ api-scanner/src/test/resources/exclude-package/java/net/corda/example/wanted/WantedClass.java
@@ -1,0 +1,7 @@
+package net.corda.example.wanted;
+
+/**
+ * This class should be included in the scan.
+ */
+public class WantedClass {
+}